Add page.querySelector(...).click() to the Script PoC (#12)
* Add page.querySelector(...).click() to the Script PoC

Add a more complex use case for the Script mock that returns a native object (PipeableElement).

During the implementation, found a way to properly use the wrappers prototypes. Refactored the code to add JS functions using the PageWrapper.prototype.

Improve how we wait for all async operations to finish. Add a wait for the whole script as well to avoid race conditions.