Add more debug info to script errors
Fix line position of errors by using an offset.
Did not find a way for async functions to output meaningful debug info as their implementation is defined in another call to `evaluateScript` and Errors get the `line` and `column` from the script in which they are created... :( Leaving it as it is for now.