PipeablePage.goto now returns a PipeableHTTPResponse (#34)
`PipeablePage.goto` now returns `PipeableHTTPStatus`, which contains the http status and the response headers.

It is captured in the Delegate and set on the PipeablePage instance. In the end decided not to bundle it with LoadPageStatus, since they have different purposes -- one tracks the load status and the other one records the http response to be returned in `goto` or in `waitForURL`.

One downside to this, actually, is that we'll be retaining the last HTTP response for the duration of the entire page, while we won't need it after the `goto` result.

Refactored the Delegate out into its own file.